+++ src/cmr/camion.c
@@ -32,7 +32,7 @@ CMR_ERROR CMRstatsCamionPrint(FILE* stream, CMR_CAMION_STATISTICS* stats, const 
     fprintf(stream, "Camion signing:\n");
     prefix = "  ";
   }
-  fprintf(stream, "%stotal: %d in %f seconds\n", prefix, stats->totalCount, stats->totalTime);
+  fprintf(stream, "%stotal: %lu in %f seconds\n", prefix, (unsigned long)stats->totalCount, stats->totalTime);
 
   return CMR_OKAY;
 }
